Unix Systems-Innovation for everyone
This is my first personal project based on HTML, CSS and JS. This website features a model name of a dummy company "Unix Systems". This website is equipped with 4 Basic web pages- Home, About Us, Services and Contact Us. Contact Us page is till in modification stage and contact us page is yet to be connected with back end. Stay tuned for the full website landing soon. Innovation for better tomorrow.